Program Runs with multiple food objects
Nullpointer exception error causd by incorrect array refercing, food.length vs food.foodCount, where foodCount is the actual number of food objects and the food.length is just the size of the array. this will be done for poison array in future updates. Also the show function for poisons need to be revamped similar to the new show function for foods.